moves LocalePicker agnostic inside its own component
